BankOCR changes
